Fix fast inverse square root to work on all systems

+++ Sensor Fusion/broadcastreceiver/madgwickahrscplusplus.cpp
@@ -225,7 +225,7 @@ void MadgwickAHRScplusplus::MadgwickAHRSupdateIMU(float gx, float gy, float gz, 
 float invSqrt(float x) {
     float halfx = 0.5f * x;
     float y = x;
-    long i = *(long*)&y;
+    int32_t i = *(int32_t*)&y;
     i = 0x5f3759df - (i>>1);
     y = *(float*)&i;
     y = y * (1.5f - (halfx * y * y));
